Moving file structure to different README.md file. Combining inscope.py into public eligibility
